moved tests to a tests directory and main functions to src
