+++ src/BijectiveHilbert.jl
@@ -1,6 +1,9 @@
 module BijectiveHilbert
 
+include("bitops.jl")
 include("hilbert.jl")
+include("hamilton.jl")
+
 export encode_hilbert_zero
 export decode_hilbert_zero
 export encode_hilbert
